refactor(chat-ws): separate v1 and v2 handlers
