+++ server/apps/api/src/routes/chat-ws/v1/index.ts
@@ -1,26 +1,26 @@
 import type Redis from 'ioredis'
 
-import type { EngagementMetrics } from '../../otel'
-import type { ChatService } from '../../services/domain/chats'
-import type { ChatWsRuntime } from './runtime'
+import type { EngagementMetrics } from '../../../otel'
+import type { ChatService } from '../../../services/domain/chats'
+import type { ChatWsRuntime } from '../runtime'
 
 import { useLogger } from '@guiiai/logg'
 import { newMessages } from '@proj-airi/server-sdk-shared/v1'
 
-import { nanoid } from '../../utils/id'
-import { createLegacyChatWsHooks } from './legacy-adapter'
-import { registerLegacyChatRpcHandlers } from './legacy-rpc'
-import { createChatWsRuntime } from './runtime'
+import { nanoid } from '../../../utils/id'
+import { createChatWsRuntime } from '../runtime'
+import { createChatWsV1Hooks } from './adapter'
+import { registerChatWsV1RpcHandlers } from './rpc'
 
 const log = useLogger('chat-ws:v1').useGlobalConfig()
 
 /**
- * Creates the legacy `/ws/chat` handlers.
+ * Creates the version-one `/ws/chat` handlers.
  *
  * The route keeps Eventa `0.3.0` and query-token authentication for clients
  * that shipped before `/ws/v2/chat`.
  */
-export function createLegacyChatWsHandlers(
+export function createChatWsV1Handlers(
   chatService: ChatService,
   redis: Redis,
   instanceId: string,
@@ -32,7 +32,7 @@ export function createLegacyChatWsHandlers(
   return function setupPeer(userId: string) {
     let connectionId: string | undefined
 
-    return createLegacyChatWsHooks({
+    return createChatWsV1Hooks({
       onContext: (ctx) => {
         const currentConnectionId = nanoid()
         connectionId = currentConnectionId
@@ -42,7 +42,7 @@ export function createLegacyChatWsHandlers(
         broadcast.ensureSubscribed(userId)
         log.withFields({ userId }).log('WS v1 connected')
 
-        registerLegacyChatRpcHandlers({
+        registerChatWsV1RpcHandlers({
           ctx,
           connectionId: currentConnectionId,
           userId,
